Add `UnimplementedHolder` (mostly like `UnknowHolder`)
This holder is meant for data that doesn't have an adequate holder yet
(where the unknown type holder is meant for data which purpose is unknown).
It behaves mostly like the unknown type holder, but also allows setting a
description string that is shown in the entry type column.
Due to issues with Visual Studio Designer I could not add a viewer tab where
it could be allowed to show the description string.
JSON stuff untested.